requirements.txt
arcgis
python-dotenv
requests
python-dateutil